feat(GiniCaptureSDK): decompose QREngagementTopView to QREngagementProgressView & QREngagementBrandView
PP-1043
